# Fixing a performance issue in SSE client