Improves the streaming RTDB API performance by checking for end of field more efficiently. Specifically, instead of matching a regex on each iteration, we keep track of the last 4 characters seen, and perform a simple string comparison.

Here are some preliminary test results to get an idea of the improvement.

| Node Size  | Time to process (Before) | Time to process (After) |
| ------------- | ------------- | ------------- |
| 100K  | 40.83s  | 1.73s |
| 1M  | - | 11.36s |

Fixes #198