@daniel-ziegler 2018-11-16
Doesn't work for me at all on Mac OS X -- I simply don't receive any listen events. This definitely needs to be tested with a real HTTP stream.

Also note that while this should hopefully improve the constant factor, it's still O(N^2).