@hiranya911 2018-11-17
There seems to be a `requests` bug here, which I've reported at https://github.com/requests/requests/issues/4876

The proposed solution essentially reduces N by reading data in chunks. With the fix I've tested payloads (in the unit tests) as large as 10MB, which can be processed under a second.