@hiranya911 2018-11-17
Attached is another solution I've been tinkering with. It doesn't change how we read content from `requests` -- just changes how we handle the incoming data:

* Using a list to buffer incoming data as opposed to string concatenation which is inefficient at large sizes.
* Getting rid of the regex check on each iteration. The alternative check I've implemented assumes '\n\n` as the only acceptable terminator, which seems to be correct for Firebase.

With this I was able to read a 1MB node from a production database in under 10 seconds. Feedback welcome.

[sse.txt](https://github.com/firebase/firebase-admin-python/files/2591340/sse.txt)