@daniel-ziegler 2018-11-17
I like that solution better since it is indeed linear time. Of course for maximum performance we'd need to figure out how to read in bigger chunks and not process each character individually in Python, but I'm happy with the improvement.