@hiranya911 2018-11-20
@daniel-ziegler I've cleaned up the 2nd solution and committed it. Can you give it a try? My performance numbers are added to the PR description.